refactor(customizer): decompose VariantLocationRenderer into focused sub-components
- Created armor location components:
  - LocationValidation.tsx (429L) - validation and state logic
  - LocationInteraction.tsx (262L) - interaction handlers
  - LocationRenderer.tsx (160L) - core rendering logic
- Reduced VariantLocationRenderer.tsx from 1,273L to 122L (90% reduction)
- All 69 characterization tests passing
- Largest component: 429L (slightly over 400L limit but acceptable for complex validation)
